requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.11.0; python_version >= '3.9'
asgiref==3.10.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.8.3; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.3; python_version >= '3.7'
cryptography==46.0.3; python_version >= '3.8' and python_full_version not in '3.9.0, 3.9.1'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distro==1.9.0; python_version >= '3.6'
dj-database-url==3.0.1
django==5.2.7; python_version >= '3.10'
django-cors-headers==4.9.0; python_version >= '3.9'
django-sendgrid-v5==1.3.0
djangorestframework==3.16.1; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
djoser==2.3.3; python_version >= '3.9' and python_version < '4.0'
google-ai-generativelanguage==0.6.15; python_version >= '3.7'
google-api-core[grpc]==2.25.1; python_version >= '3.7'
google-api-python-client==2.183.0; python_version >= '3.7'
google-auth==2.40.3; python_version >= '3.7'
google-auth-httplib2==0.2.0
google-generativeai==0.8.5; python_version >= '3.9'
googleapis-common-protos==1.70.0; python_version >= '3.7'
grpcio==1.75.0; python_version >= '3.9'
grpcio-status==1.71.2; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
httpcore==1.0.9; python_version >= '3.8'
httplib2==0.31.0; python_version >= '3.6'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jiter==0.11.0; python_version >= '3.9'
markupsafe==3.0.3; python_version >= '3.9'
oauthlib==3.3.1; python_version >= '3.8'
openai==1.109.1; python_version >= '3.8'
proto-plus==1.26.1; python_version >= '3.7'
protobuf==5.29.5; python_version >= '3.8'
psycopg2==2.9.10; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pyasn1-modules==0.4.2; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pydantic==2.11.9; python_version >= '3.9'
pydantic-core==2.33.2; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
pyparsing==3.2.5; python_version >= '3.9'
python-http-client==3.3.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python3-openid==3.2.0
requests==2.32.5; python_version >= '3.9'
requests-oauthlib==2.0.0; python_version >= '3.4'
rsa==4.9.1; python_version >= '3.6' and python_version < '4'
sendgrid==6.12.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
sniffio==1.3.1; python_version >= '3.7'
social-auth-app-django==5.5.1; python_version >= '3.10'
social-auth-core==4.7.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.1; python_version >= '3.9'
uritemplate==4.2.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
